+++ web/src/component.dart
@@ -48,3 +48,4 @@ part 'component/widget/file_upload_model.dart';
 part 'component/widget/hidden_model.dart';
 part 'component/widget/html_model.dart';
 part 'component/widget/hyperlink_model.dart';
+part 'component/widget/image_model.dart';